Moved path mapping support to MayaUsdProxyShapeSceneIndexBase.
